Update: kafka and mongodb charts deployment; model Movie, AggregateRating

+++ model/aggregate_rating.go
@@ -5,5 +5,5 @@ type AggregateRating struct {
 	RatingCount int     `json:"ratingCount,omitempty"`
 	BestRating  int     `json:"bestRating,omitempty"`
 	WorstRating int     `json:"worstRating,omitempty"`
-	RatingValue float32 `json:"ratingValue,omitempty"`
+	RatingValue float64 `json:"ratingValue,omitempty"`
 }
